[simulation] SWORD-645 Evolution des réseaux souterrains (simulation) + Nettoyage

+++ src/libraries/simulation_kernel/Entities/Agents/Actions/Objects/PHY_ActionDistributeObject.cpp
@@ -18,7 +18,7 @@
 // Name: PHY_ActionDistributeObject constructor
 // Mined: NLD 2004-08-18
 // -----------------------------------------------------------------------------
-PHY_ActionDistributeObject::PHY_ActionDistributeObject( MIL_AgentPion& pion, boost::shared_ptr< DEC_Knowledge_Object > pObject, int quantity )
+PHY_ActionDistributeObject::PHY_ActionDistributeObject( MIL_AgentPion& pion, boost::shared_ptr< DEC_Knowledge_Object > pObject, double quantity )
     : PHY_DecisionCallbackAction_ABC( pion )
     , role_    ( pion.GetRole< PHY_RoleAction_Objects >() )
     , pObject_ ( pObject )
